Renamed TripModeType
