+++ src/app/shared/ojp-sdk/index.ts
@@ -13,4 +13,4 @@ export * from './types/geo-restriction.type'
 export * from './types/individual-mode.types'
 export * from './types/journey-points'
 export * from './types/stage-config'
-export * from './types/trip-mot-type'
+export * from './types/trip-mode-type'
