+++ src/app/shared/ojp-sdk/request/journey-request/journey-request-params.ts
@@ -1,7 +1,7 @@
 import { Location } from "../../location/location"
 import { TripLocationPoint } from "../../trip";
 import { IndividualTransportMode } from "../../types/individual-mode.types";
-import { TripModeType } from "../../types/trip-mot-type";
+import { TripModeType } from "../../types/trip-mode-type";
 
 export class JourneyRequestParams {
   tripLocations: TripLocationPoint[]
