+++ src/app/shared/ojp-sdk/types/trip-mode-type.ts
